+++ doc/source/releasenotes.rst
@@ -5,5 +5,8 @@ Release Notes
 -----
 
 * Added capability to add comments
+* Added extension for build results
+* Added :py:meth:`osctiny.packages.Package.checkout` to properly check-out an
+  entire package
 * On initialization :py:class:`osctiny.Osc` accepts a ``cache`` keyword argument
   to sache API responses in a cache dictionary.